Refactor workspace configuration in sway and i3
This commit refactors the workspace configuration in sway and i3. In sway, it adds a new workspace with the number "1" and includes the `/etc/sway/config.d/*` file and the `~/.config/sway/startup.conf` and `~/.config/sway/keybinds.conf` files. In i3, it updates the `startups.conf` file to include the `playerctld daemon` command.